take 6

+++ src/firefly_base/firefly_base/gui/components/FBMSEGEditor.cpp
@@ -7,7 +7,7 @@ FBMSEGEditor(
   int maxPoints,
   int maxLengthRatioNum,
   int maxLengthRatioDen,
-  float maxLengthReal,
+  double maxLengthReal,
   int gridMinRatioGranularity):
 Component(),
 _maxPoints(maxPoints),
